correct bug with confusion matrix (if split<100%)

+++ dzetsaka.py
@@ -1087,7 +1087,7 @@ class dzetsaka ( QDialog ):
                     inSeed = 0
                     if self.dockwidget.checkOutMatrix.isChecked():
                         outMatrix = self.dockwidget.outMatrix.text()
-                        inSplit = self.dockwidget.inSplit.value()/100
+                        inSplit = float(self.dockwidget.inSplit.value()/100.)
                     else:
                         inSplit = 1
                         outMatrix = None
